Add office user show component
