+++ src/scenes/SystemAdmin/AdminPagination.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { Pagination } from 'react-admin';
+
+const AdminPagination = props => <Pagination rowsPerPageOptions={[]} {...props} />;
+
+export default AdminPagination;
